execution-engine: rename/restructure engine module
