+++ execution-engine/comm/src/engine_server/mod.rs
@@ -4,7 +4,8 @@ use std::fmt::Debug;
 use std::marker::{Send, Sync};
 
 use common::key::Key;
-use execution_engine::engine::{EngineState, Error as EngineError};
+use execution_engine::engine_state::error::Error as EngineError;
+use execution_engine::engine_state::EngineState;
 use execution_engine::execution::{Executor, WasmiExecutor};
 use execution_engine::tracking_copy::QueryResult;
 use ipc::*;
